move cypress tests to tests/cypress/integration
