+++ tests/cypress/cypress.json
@@ -1,5 +1,5 @@
 {
-  "integrationFolder": "tests/cypress/",
+  "integrationFolder": "tests/cypress/integration",
   "video": false,
   "pluginsFile": "tests/cypress/plugins/index.js",
   "supportFile": "tests/cypress/support/index.js",
